chore: Rename 'definitionLogic.ts' file to 'symbolLocation.ts' and update module import in 'handlersWrapper.ts'
